requirements.txt
azure-cognitiveservices-speech
discord.py
PyNaCl
